bot/requirements.txt
aiogram>=3.0.0
aiohttp==3.9.3
asyncpg
requests==2.31.0
python-dotenv==1.1.0
sqlalchemy
uvicorn
